+++ noodles-bam/examples/bam_view.rs
@@ -5,7 +5,7 @@
 use std::{env, io};
 
 use noodles_bam as bam;
-use noodles_sam::{self as sam, AlignmentWriter};
+use noodles_sam::{self as sam, alignment::io::Writer};
 
 fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");