+++ noodles-sam/src/alignment/io.rs
@@ -1,5 +1,6 @@
 //! Alignment record I/O.
 
 mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};