+++ noodles-sam/src/alignment/io/writer.rs
@@ -1,11 +1,11 @@
 use std::io;
 
-use super::{alignment::Record, Header};
+use crate::{alignment::Record, Header};
 
 /// An alignment format writer.
 ///
 /// A call to [`Self::finish`] must be made before the writer is dropped.
-pub trait AlignmentWriter {
+pub trait Writer {
     /// Writes a SAM header.
     fn write_alignment_header(&mut self, header: &Header) -> io::Result<()>;